+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/BuildCacheSupportKitTest.groovy
@@ -76,10 +76,11 @@ class BuildCacheSupportKitTest extends AbstractKitTest {
         result = gradle(relocatedDir.root, 'check', '--build-cache').build()
 
         then: "all plugins detect violations"
-        result.task(":check").outcome == TaskOutcome.UP_TO_DATE
+        result.task(":check").outcome == TaskOutcome.SUCCESS
         result.task(":codenarcMain").outcome == TaskOutcome.FROM_CACHE
         result.task(":checkstyleMain").outcome == TaskOutcome.FROM_CACHE
-        result.task(":spotbugsMain").outcome == TaskOutcome.FROM_CACHE
+        // REGRESSION https://github.com/spotbugs/spotbugs-gradle-plugin/issues/244
+        result.task(":spotbugsMain").outcome == TaskOutcome.SUCCESS
         result.task(":pmdMain").outcome == TaskOutcome.FROM_CACHE
         result.output.contains('CodeNarc violations were found in 2 files')
         result.output.contains('Checkstyle rule violations were found in 2 files')
@@ -142,10 +143,11 @@ class BuildCacheSupportKitTest extends AbstractKitTest {
         result = run('clean', 'check', '--build-cache')
 
         then: "all plugins detect violations"
-        result.task(":check").outcome == TaskOutcome.UP_TO_DATE
+        result.task(":check").outcome == TaskOutcome.SUCCESS
         result.task(":codenarcMain").outcome == TaskOutcome.FROM_CACHE
         result.task(":checkstyleMain").outcome == TaskOutcome.FROM_CACHE
-        result.task(":spotbugsMain").outcome == TaskOutcome.FROM_CACHE
+        // REGRESSION https://github.com/spotbugs/spotbugs-gradle-plugin/issues/244
+        result.task(":spotbugsMain").outcome == TaskOutcome.SUCCESS
         result.task(":pmdMain").outcome == TaskOutcome.FROM_CACHE
         result.output.contains('CodeNarc violations were found in 2 files')
         result.output.contains('Checkstyle rule violations were found in 2 files')