+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/HtmlReportsDisableKitTest.groovy
@@ -43,7 +43,7 @@ class HtmlReportsDisableKitTest extends AbstractKitTest {
         result.task(":check").outcome == TaskOutcome.SUCCESS
         result.output.contains('CodeNarc rule violations were found')
         result.output.contains('Checkstyle rule violations were found')
-        result.output.contains('SpotBugs rule violations were found')
+        result.output.contains('SpotBugs reported failures')
         result.output.contains('PMD rule violations were found')
 
         then: "html reports are not generated"