+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/QualityPluginTest.groovy
@@ -1,6 +1,6 @@
 package ru.vyarus.gradle.plugin.quality
 
-import com.github.spotbugs.SpotBugsPlugin
+import com.github.spotbugs.snom.SpotBugsBasePlugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.quality.CheckstylePlugin
 import org.gradle.api.plugins.quality.CodeNarcPlugin
@@ -25,7 +25,7 @@ class QualityPluginTest extends AbstractTest {
         then: "plugins registered"
         project.plugins.findPlugin(CheckstylePlugin)
         project.plugins.findPlugin(PmdPlugin)
-        project.plugins.findPlugin(SpotBugsPlugin)
+        project.plugins.findPlugin(SpotBugsBasePlugin)
         !project.plugins.findPlugin(CodeNarcPlugin)
 
         then: "tasks installed"
@@ -47,7 +47,7 @@ class QualityPluginTest extends AbstractTest {
         then: "plugins registered"
         project.plugins.findPlugin(CheckstylePlugin)
         project.plugins.findPlugin(PmdPlugin)
-        project.plugins.findPlugin(SpotBugsPlugin)
+        project.plugins.findPlugin(SpotBugsBasePlugin)
         !project.plugins.findPlugin(CodeNarcPlugin)
 
         then: "tasks installed"
@@ -69,7 +69,7 @@ class QualityPluginTest extends AbstractTest {
         then: "plugins registered"
         !project.plugins.findPlugin(CheckstylePlugin)
         !project.plugins.findPlugin(PmdPlugin)
-        !project.plugins.findPlugin(SpotBugsPlugin)
+        !project.plugins.findPlugin(SpotBugsBasePlugin)
         project.plugins.findPlugin(CodeNarcPlugin)
     }
 
@@ -87,7 +87,7 @@ class QualityPluginTest extends AbstractTest {
         then: "plugins registered"
         !project.plugins.findPlugin(CheckstylePlugin)
         !project.plugins.findPlugin(PmdPlugin)
-        !project.plugins.findPlugin(SpotBugsPlugin)
+        !project.plugins.findPlugin(SpotBugsBasePlugin)
         project.plugins.findPlugin(CodeNarcPlugin)
     }
 
@@ -111,7 +111,7 @@ class QualityPluginTest extends AbstractTest {
         then: "plugins not registered"
         !project.plugins.findPlugin(CheckstylePlugin)
         !project.plugins.findPlugin(PmdPlugin)
-        !project.plugins.findPlugin(SpotBugsPlugin)
+        !project.plugins.findPlugin(SpotBugsBasePlugin)
         !project.plugins.findPlugin(CodeNarcPlugin)
     }
 
@@ -165,7 +165,7 @@ class QualityPluginTest extends AbstractTest {
         then: "plugins registered"
         project.plugins.findPlugin(CheckstylePlugin)
         !project.plugins.findPlugin(PmdPlugin)
-        !project.plugins.findPlugin(SpotBugsPlugin)
+        !project.plugins.findPlugin(SpotBugsBasePlugin)
         !project.plugins.findPlugin(CodeNarcPlugin)
     }
 }