+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+* Update spotbugs plugin to 4.0.8 (#26)
+    - Fixes gradle 6.4 compatibility.
+    - The new plugin is a re-write of the original spotbugs plugin. But quality plugin 
+      makes it work as before, so you shouldn't see any difference.
+    - Minimal supported gradle is now 5.6 (due to spotbugs plugin)  
+    - New spotbugs plugin can generate html report itself, but this option is not used
+
+WARNING:
+    * New spotbugs plugin does not support build cache (https://github.com/spotbugs/spotbugs-gradle-plugin/issues/244)
+    * Spotbugs task always show an exception when violations found (not a problem, just confusing)
+
 ### 4.2.2 (2020-04-23)
 * Remove spotbugs configuration "tuning" (added in the last version) because it eventually works incorrectly (often causing warnings).
   Spotbugs 4.0.2 depends on slf4j 1.8 which should avoid compatibility problems (at least for some time).